VoiceAssistant/requirements.txt
Flask
Flask_Cors
ollama
requests
pyttsx3
openai-whisper
dotenv
coqui-tts
coqui-tts[codec]
torchaudio
suno-bark @ git+https://github.com/suno-ai/bark.git@f4f32d4cd480dfec1c245d258174bc9bde3c2148